+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -5,8 +5,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.auth.CheckRoleInterceptor;
-import roomescape.auth.LoginMemberArgumentResolver;
+import roomescape.controller.auth.CheckRoleInterceptor;
+import roomescape.controller.auth.LoginMemberArgumentResolver;
 import roomescape.service.MemberService;
 
 import java.util.List;
