+++ src/main/java/roomescape/controller/auth/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.controller.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
