+++ src/main/java/roomescape/controller/auth/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.controller.auth;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
