+++ src/main/java/roomescape/controller/auth/CheckRoleInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.controller.auth;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
