+++ src/main/java/roomescape/controller/reservation/AdminReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.reservation;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
